+++ app/src/main/kotlin/diettracker/Routing.kt
@@ -1,10 +1,5 @@
 package diettracker
 
-import diettracker.db.tables.Professionals
-import diettracker.db.tables.Recipes
-import diettracker.db.tables.Roles
-import diettracker.db.tables.UserRoles
-import diettracker.db.tables.Users
 import io.ktor.server.application.Application
 import io.ktor.server.auth.authenticate
 import io.ktor.server.http.content.staticResources
@@ -18,9 +13,8 @@ import io.ktor.server.routing.post
 import io.ktor.server.routing.routing
 import io.ktor.server.sessions.get
 import io.ktor.server.sessions.sessions
-import org.jetbrains.exposed.v1.core.eq
-import org.jetbrains.exposed.v1.jdbc.selectAll
-import org.jetbrains.exposed.v1.jdbc.transactions.transaction
+
+private const val DEFAULT_GRAMS = 100
 
 fun Application.configureRouting() {
     routing {
@@ -59,6 +53,28 @@ fun Route.configurePublicRoutes() {
         )
     }
 
+    configureFoodRoutes()
+
+    authenticate("group49-client_auth") {
+        get("/") { call.dashboardPage() }
+        get("/logout") { call.logout() }
+    }
+
+    get("/diary") {
+        call.respond(PebbleContent("pages/client_dash/food_diary.peb", mapOf("showNavbar" to true)))
+    }
+
+    get("/recipes") {
+        val recipes = getAllRecipes()
+        call.respond(PebbleContent("pages/recipes/recipes.peb", mapOf("showNavbar" to true, "recipes" to recipes)))
+    }
+
+    get("/health") {
+        call.respondText("OK")
+    }
+}
+
+fun Route.configureFoodRoutes() {
     get("/food_log") {
         val recipeQuery = call.request.queryParameters["query"]
         val foodQuery = call.request.queryParameters["foodquery"]
@@ -68,23 +84,24 @@ fun Route.configurePublicRoutes() {
             val recipes = searchRecipes(recipeQuery)
             call.respondTemplate(
                 "pages/client_dash/add_food.peb",
-                mapOf("recipes" to recipes, "calories" to calories)
+                mapOf("recipes" to recipes, "calories" to calories),
             )
         } else if (foodQuery != null && foodQuery.isNotBlank()) {
             val foods = searchFoods(foodQuery)
             call.respondTemplate(
                 "pages/client_dash/add_food.peb",
-                mapOf("foods" to foods, "calories" to calories)
+                mapOf("foods" to foods, "calories" to calories),
             )
         } else {
             call.foodLogPage()
         }
     }
-    post("/food_log_recipe"){
+
+    post("/food_log_recipe") {
         call.foodLogRecipe()
     }
 
-    post("/food_log_custom"){
+    post("/food_log_custom") {
         call.foodLogCustom()
     }
 
@@ -102,71 +119,19 @@ fun Route.configurePublicRoutes() {
     get("/food_search") {
         val query = call.request.queryParameters["foodquery"] ?: ""
         val foods = searchFoods(query)
-        val grams = call.request.queryParameters["grams"]?.toIntOrNull() ?: 100
+        val grams = call.request.queryParameters["grams"]?.toIntOrNull() ?: DEFAULT_GRAMS
         val calories = call.sessions.get<CaloriesSession>()?.calories ?: 0
 
-        call.respondTemplate("pages/client_dash/add_food.peb", mapOf("foods" to foods, "calories" to calories))
-    }
-
-
-    authenticate("group49-client_auth") {
-        get("/") { call.dashboardPage() }
-        get("/logout") { call.logout() }
-    }
-
-
-    get("/diary") {
-        call.respond(PebbleContent("pages/client_dash/food_diary.peb", mapOf("showNavbar" to true)))
-    }
-
-    get("/recipes") {
-        val recipes = getAllRecipes()
-        call.respond(PebbleContent("pages/recipes/recipes.peb", mapOf("showNavbar" to true, "recipes" to recipes)))
-    }
-
-    get("/health") {
-        call.respondText("OK")
+        call.respondTemplate(
+            "pages/client_dash/add_food.peb",
+            mapOf(
+                "foods" to foods,
+                "calories" to calories,
+                "grams" to grams,
+            ),
+        )
     }
 }
-
-fun getUserIdByEmail(email: String): Int? =
-    transaction {
-        Users.selectAll()
-            .where { Users.email eq email }
-            .map { it[Users.user_id] }
-            .singleOrNull()
-    }
-
-fun getUserRoles(userId: Int): List<String> =
-    transaction {
-        (UserRoles innerJoin Roles)
-            .selectAll()
-            .where { UserRoles.user_id eq userId }
-            .map { it[Roles.role_name] }
-    }
-
-fun getAllRecipes(): List<Map<String, Any?>> =
-    transaction {
-        Recipes.selectAll().map { row ->
-            mapOf(
-                "id" to row[Recipes.recipes_id],
-                "name" to row[Recipes.recipe_name],
-                "thumbnail" to row[Recipes.thumbnail_url],
-            )
-        }
-    }
-
-fun getAllProfessionals(): List<Map<String, Any>> =
-    transaction {
-        Professionals.selectAll().map { row ->
-            mapOf(
-                "id" to row[Professionals.professional_id],
-                "job_title" to row[Professionals.job_title],
-                "organisation" to row[Professionals.organistation],
-                "bio" to row[Professionals.bio],
-            )
-        }
-    }
 
 fun Route.configureAuthRoutes() {
     get("/Sign-Up") { call.signUpPage() }
