+++ app/src/main/kotlin/diettracker/UserDatabase.kt
@@ -1,5 +1,7 @@
 package diettracker
 
+import diettracker.db.tables.Professionals
+import diettracker.db.tables.Recipes
 import diettracker.db.tables.Roles
 import diettracker.db.tables.UserRoles
 import diettracker.db.tables.Users
@@ -9,6 +11,45 @@ import org.jetbrains.exposed.v1.jdbc.selectAll
 import org.jetbrains.exposed.v1.jdbc.transactions.transaction
 import org.mindrot.jbcrypt.BCrypt
 import java.time.Instant
+
+fun getUserIdByEmail(email: String): Int? =
+    transaction {
+        Users.selectAll()
+            .where { Users.email eq email }
+            .map { it[Users.user_id] }
+            .singleOrNull()
+    }
+
+fun getUserRoles(userId: Int): List<String> =
+    transaction {
+        (UserRoles innerJoin Roles)
+            .selectAll()
+            .where { UserRoles.user_id eq userId }
+            .map { it[Roles.role_name] }
+    }
+
+fun getAllRecipes(): List<Map<String, Any?>> =
+    transaction {
+        Recipes.selectAll().map { row ->
+            mapOf(
+                "id" to row[Recipes.recipes_id],
+                "name" to row[Recipes.recipe_name],
+                "thumbnail" to row[Recipes.thumbnail_url],
+            )
+        }
+    }
+
+fun getAllProfessionals(): List<Map<String, Any>> =
+    transaction {
+        Professionals.selectAll().map { row ->
+            mapOf(
+                "id" to row[Professionals.professional_id],
+                "job_title" to row[Professionals.job_title],
+                "organisation" to row[Professionals.organistation],
+                "bio" to row[Professionals.bio],
+            )
+        }
+    }
 
 object UserDatabase {
     fun checkCreds(
@@ -52,18 +93,20 @@ object UserDatabase {
 
             if (exists) return@transaction false
 
-            val newUserId = Users.insert {
-                it[Users.first_name] = normalisedEmail.substringBefore("@")
-                it[Users.second_name] = ""
-                it[Users.email] = normalisedEmail
-                it[Users.password_hash] = passwordHash
-                it[Users.created_at] = Instant.now()
-            } get Users.user_id
+            val newUserId =
+                Users.insert {
+                    it[Users.first_name] = normalisedEmail.substringBefore("@")
+                    it[Users.second_name] = ""
+                    it[Users.email] = normalisedEmail
+                    it[Users.password_hash] = passwordHash
+                    it[Users.created_at] = Instant.now()
+                } get Users.user_id
 
-            val clientRoleId = Roles.selectAll()
-                .where { Roles.role_name eq "client" }
-                .map { it[Roles.role_id] }
-                .singleOrNull()
+            val clientRoleId =
+                Roles.selectAll()
+                    .where { Roles.role_name eq "client" }
+                    .map { it[Roles.role_id] }
+                    .singleOrNull()
 
             if (clientRoleId != null) {
                 UserRoles.insert {
